pyproject.toml
[tool.poetry]
name = "api_prestcontas"
version = "0.1.0"
description = ""
authors = ["ReinaldoJuvino <reinaldojuvinoneto@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.1.3"
requests = "^2.24.0"
django-ckeditor = "^6.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
